feat: refactor generation context, channels, and app metadata
