+++ README.md
@@ -4,7 +4,7 @@ GenStudio UI Extensibility SDK
 **This repo is still WIP. Please use caution and reach out to GS team before using**
 
 ## API Documentation
-[API Doc](./doc/api.md)
+[API Doc](./docs/README.md)
 
 ## Usage
 ```
