+++ docs/README.md
@@ -8,7 +8,7 @@ GenStudio UI Extensibility SDK
 **This repo is still WIP. Please use caution and reach out to GS team before using**
 
 ## API Documentation
-[API Doc](./doc/api.md)
+[API Doc](./docs/README.md)
 
 ## Usage
 ```
