+++ docs/globals.md
@@ -8,6 +8,16 @@
 
 - [AdditionalContextTypes](enumerations/AdditionalContextTypes.md)
 
+## Classes
+
+- [ExperienceError](classes/ExperienceError.md)
+- [ExperienceService](classes/ExperienceService.md)
+
+## Interfaces
+
+- [Experience](interfaces/Experience.md)
+- [ExperienceField](interfaces/ExperienceField.md)
+
 ## Type Aliases
 
 - [AdditionalContext](type-aliases/AdditionalContext.md)
@@ -16,14 +26,7 @@
 - [Brand](type-aliases/Brand.md)
 - [Channel](type-aliases/Channel.md)
 - [Claim](type-aliases/Claim.md)
-- [Experience](type-aliases/Experience.md)
-- [ExperienceField](type-aliases/ExperienceField.md)
-- [FieldRole](type-aliases/FieldRole.md)
 - [GenerationContext](type-aliases/GenerationContext.md)
 - [Persona](type-aliases/Persona.md)
 - [Product](type-aliases/Product.md)
 - [SectionGenerationContext](type-aliases/SectionGenerationContext.md)
-
-## Variables
-
-- [Email](variables/Email.md)
