+++ docs/interfaces/Experience.md
@@ -1,0 +1,26 @@
+[**@adobe/genstudio-uix-sdk**](../README.md)
+
+***
+
+[@adobe/genstudio-uix-sdk](../globals.md) / Experience
+
+# Interface: Experience
+
+Represents an Experience entity in the system.
+An Experience is a container for various fields that define its characteristics.
+
+## Properties
+
+### experienceFields
+
+> **experienceFields**: `Record`\<`string`, [`ExperienceField`](ExperienceField.md)\>
+
+Collection of experience fields stored as key-value pairs
+
+***
+
+### id
+
+> **id**: `string`
+
+Unique identifier for the experience
