+++ docs/interfaces/ExperienceField.md
@@ -1,0 +1,26 @@
+[**@adobe/genstudio-uix-sdk**](../README.md)
+
+***
+
+[@adobe/genstudio-uix-sdk](../globals.md) / ExperienceField
+
+# Interface: ExperienceField
+
+Represents a field within an Experience.
+Each field contains a name and corresponding value.
+
+## Properties
+
+### fieldName
+
+> **fieldName**: `string`
+
+Name of the experience field
+
+***
+
+### fieldValue
+
+> **fieldValue**: `string`
+
+Value associated with the experience field
