--- docs/type-aliases/Experience.md
@@ -1,23 +1,0 @@
-[**@adobe/genstudio-uix-sdk**](../README.md)
-
-***
-
-[@adobe/genstudio-uix-sdk](../globals.md) / Experience
-
-# Type Alias: Experience
-
-> **Experience**: \{ `experienceFields`: \{\}; `id`: `string`; \}
-
-## Type declaration
-
-### experienceFields
-
-> **experienceFields**: \{\}
-
-#### Index Signature
-
-\[`key`: `string`\]: [`ExperienceField`](ExperienceField.md)
-
-### id
-
-> **id**: `string`
