--- docs/type-aliases/ExperienceField.md
@@ -1,27 +1,0 @@
-[**@adobe/genstudio-uix-sdk**](../README.md)
-
-***
-
-[@adobe/genstudio-uix-sdk](../globals.md) / ExperienceField
-
-# Type Alias: ExperienceField
-
-> **ExperienceField**: \{ `fieldName`: `string`; `fieldRole`: [`FieldRole`](FieldRole.md); `fieldValue`: `string` \| `Object`; `readonly`: `boolean`; \}
-
-## Type declaration
-
-### fieldName
-
-> **fieldName**: `string`
-
-### fieldRole
-
-> **fieldRole**: [`FieldRole`](FieldRole.md)
-
-### fieldValue
-
-> **fieldValue**: `string` \| `Object`
-
-### readonly
-
-> **readonly**: `boolean`
