--- docs/type-aliases/FieldRole.md
@@ -1,19 +1,0 @@
-[**@adobe/genstudio-uix-sdk**](../README.md)
-
-***
-
-[@adobe/genstudio-uix-sdk](../globals.md) / FieldRole
-
-# Type Alias: FieldRole
-
-> **FieldRole**: \{ `name`: `string`; \}
-
-Represents the role of a field within an experience.
-
-## Type declaration
-
-### name
-
-> **name**: `string`
-
-The name of the field role.
