+++ docs/type-aliases/SectionGenerationContext.md
@@ -6,13 +6,13 @@
 
 # Type Alias: SectionGenerationContext
 
-> **SectionGenerationContext**: \{ `additionalContexts`: `Map`\<`string`, [`AdditionalContext`](AdditionalContext.md)\<`any`\>\>; `id`: `string`; `product`: [`Product`](Product.md); \}
+> **SectionGenerationContext**: \{ `additionalContexts`: `Record`\<`string`, [`AdditionalContext`](AdditionalContext.md)\<`any`\>\>; `id`: `string`; `product`: [`Product`](Product.md); \}
 
 ## Type declaration
 
 ### additionalContexts
 
-> **additionalContexts**: `Map`\<`string`, [`AdditionalContext`](AdditionalContext.md)\<`any`\>\>
+> **additionalContexts**: `Record`\<`string`, [`AdditionalContext`](AdditionalContext.md)\<`any`\>\>
 
 ### id
