--- docs/variables/Email.md
@@ -1,11 +1,0 @@
-[**@adobe/genstudio-uix-sdk**](../README.md)
-
-***
-
-[@adobe/genstudio-uix-sdk](../globals.md) / Email
-
-# Variable: Email
-
-> `const` **Email**: [`Channel`](../type-aliases/Channel.md)
-
-Generation Context
