+++ tests/types/channel/Channel.test.ts
@@ -10,11 +10,16 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-export * from "./types/appMetaData/AppMetaData";
-export * from "./types/channel/Channel";
-export * from "./types/generationContext/GenerationContext";
-export * from "./types/experience/Experience";
-export {
-  ExperienceService,
-  ExperienceError,
-} from "./types/experience/ExperienceService";
+import { Channel } from "../../../src/types/channel/Channel";
+
+describe("contract", () => {
+  it("should define Channel", () => {
+    const channel: Channel = {
+      id: "Email",
+      name: "Email",
+    };
+    expect(channel).toBeDefined();
+    expect(channel.id).toBe("Email");
+    expect(channel.name).toBe("Email");
+  });
+});
